sed: add --error-if-unmodified to error out if input was not mutated
Gentoo makes heavy uses of sed in its ebuilds [1]. However, there is
currently no reliable, robust nor easy way to detect if a sed call is
still required, i.e., if it still mutates the input.

Previous attempts in Gentoo to remedy the situation involved
more-or-less heavy bash machinery to detect redundant/unnecessary
invocations of sed that became no-ops [2]. In the discussions around
this, it was pointed out that the right place to fix this, is, indeed,
sed.

Hence this commit introduces the --error-if-unmodified flag, which, if
enabled, will return an exit status of 8 in case sed did not apply any
modifications to the input.

This will greatly help Gentoo with identifying stale sed invocations,
and I believe others are in the same situation and would benefit form
it too.

1: https://devmanual.gentoo.org/tools-reference/sed/
2: https://archives.gentoo.org/gentoo-dev/message/29cdbda90ea5b8127eee73a3fd036037